refactor(app): move di to app file
