+++ lib/app/view/app.dart
@@ -1,45 +1,56 @@
 import 'package:app_ui/app_ui.dart';
+import 'package:blueprint/app/dependency_injection/init.dart';
 import 'package:blueprint/app/routes/routes.dart';
 import 'package:blueprint/authentication/state_management/authentication_cubit/authentication_cubit.dart';
+import 'package:blueprint/blueprint/state_management/todays_blueprint/todays_blueprint_cubit.dart';
 import 'package:blueprint/core/l10n/l10n.dart';
+import 'package:blueprint/integrations/state_management/integrations_repository/integrations_cubit.dart';
 import 'package:blueprint/settings/state_management/bloc/settings_bloc.dart';
+import 'package:blueprint/tasks/state_management/cubit/tasks_cubit.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 
-class App extends StatefulWidget {
+class App extends StatelessWidget {
   const App({super.key});
 
   @override
-  State<App> createState() => _AppState();
+  Widget build(BuildContext context) {
+    return MultiBlocProvider(
+      providers: [
+        BlocProvider.value(
+          value: sl<AuthenticationCubit>(),
+        ),
+        BlocProvider.value(value: sl<SettingsBloc>()),
+        BlocProvider.value(value: sl<TasksCubit>()),
+        BlocProvider.value(value: sl<IntegrationsCubit>()),
+        BlocProvider.value(value: sl<TodaysBlueprintCubit>()),
+      ],
+      child: const AppView(),
+    );
+  }
 }
 
-class _AppState extends State<App> {
-  late AppRouter appRouter;
-
-  @override
-  void initState() {
-    appRouter = AppRouter(
-      authenticationCubit: context.read<AuthenticationCubit>(),
-    );
-    super.initState();
-  }
+class AppView extends StatelessWidget {
+  const AppView({super.key});
 
   @override
   Widget build(BuildContext context) {
+    final appRouter = AppRouter(
+      authenticationCubit: context.read<AuthenticationCubit>(),
+    );
+
     return Builder(
       builder: (context) {
-        return BlocSelector<SettingsBloc, SettingsState, AppBrightness>(
-          selector: (state) => state.brightness,
-          builder: (context, brightness) {
-            return MaterialApp.router(
-              theme: lightTheme,
-              themeMode: brightness.themeMode,
-              darkTheme: darkTheme,
-              localizationsDelegates: AppLocalizations.localizationsDelegates,
-              supportedLocales: AppLocalizations.supportedLocales,
-              routerConfig: appRouter.config(),
-            );
-          },
+        final brightness =
+            context.select((SettingsBloc bloc) => bloc.state.brightness);
+
+        return MaterialApp.router(
+          theme: lightTheme,
+          themeMode: brightness.themeMode,
+          darkTheme: darkTheme,
+          localizationsDelegates: AppLocalizations.localizationsDelegates,
+          supportedLocales: AppLocalizations.supportedLocales,
+          routerConfig: appRouter.config(),
         );
       },
     );
