+++ lib/main_development.dart
@@ -1,5 +1,4 @@
 import 'package:blueprint/app/app.dart';
-import 'package:blueprint/app/dependency_injection/global_dependencies_provider.dart';
 import 'package:blueprint/bootstrap.dart';
 import 'package:blueprint/core/env/app_environment.dart';
 import 'package:blueprint/core/firebase_options/firebase_options_dev.dart';
@@ -12,9 +11,7 @@ void main() async {
     options: DefaultFirebaseOptions.currentPlatform,
   );
   await bootstrap(
-    () => const GlobalDependenciesProvider(
-      child: App(),
-    ),
+    () => const App(),
     environment: AppEnvironment.development,
   );
 }
