+++ lib/main_staging.dart
@@ -1,5 +1,5 @@
 import 'package:blueprint/app/app.dart';
-import 'package:blueprint/app/dependency_injection/global_dependencies_provider.dart';
+
 import 'package:blueprint/bootstrap.dart';
 import 'package:blueprint/core/env/app_environment.dart';
 import 'package:blueprint/core/firebase_options/firebase_options_stg.dart';
@@ -13,9 +13,7 @@ void main() async {
   );
 
   await bootstrap(
-    () => const GlobalDependenciesProvider(
-      child: App(),
-    ),
+    () => const App(),
     environment: AppEnvironment.staging,
   );
 }
